(struct direct): Use `unsigned int' instead of `unsigned long int' for `d_fileno' member.

+++ sysdeps/unix/bsd/direct.h
@@ -4,9 +4,9 @@
 
 struct direct
   {
-    unsigned long int d_fileno;
-    unsigned short int d_reclen;
-    unsigned short int d_namlen;
+    unsigned int d_fileno;	/* 32 bits.  */
+    unsigned short int d_reclen; /* 16 bits.  */
+    unsigned short int d_namlen; /* 16 bits.  */
     char d_name[MAXNAMLEN + 1];
   };
